+++ azext_iot/deviceupdate/command_map.py
@@ -110,6 +110,7 @@ def load_deviceupdate_commands(self, _):
         command_type=deviceupdate_update_ops,
     ) as cmd_group:
         cmd_group.command("v5", "manifest_init_v5", is_experimental=True)
+        cmd_group.show_command("calculate-hash", "calculate_hash")
 
     with self.command_group(
         "iot device-update device",