+++ azext_iot/deviceupdate/params.py
@@ -19,6 +19,7 @@ from azext_iot.deviceupdate.common import (
     ADUPrivateLinkServiceConnectionStatus,
     ADUAccountSKUType,
     ADUManageDeviceImportType,
+    ADUValidHashAlgorithmType,
 )
 
 
@@ -474,7 +475,7 @@ def load_deviceupdate_arguments(self, _):
             nargs="+",
             action="append",
             help="Space-separated key=value pairs corresponding to properties of a device this update is compatible with. "
-            "Typically used for defining properties such as deviceManufacturer and deviceModel. "
+            "Typically used for defining properties such as manufacturer and model. "
             "--compat can be used 1 or more times. ",
         )
         context.argument(
@@ -490,7 +491,7 @@ def load_deviceupdate_arguments(self, _):
             "`updateId.provider`, `updateId.name` `updateId.version` and `description`."
             "The following inline step keys are supported: "
             "`handler` (ex: 'microsoft/script:1' or 'microsoft/swupdate:1' or 'microsoft/apt:1'), "
-            "`properties` (in-line json object the agent will pass to the handler) and `description`. "
+            "`properties` (in-line json object the agent will pass to the handler), and `description`. "
             "--step can be used 1 or more times.",
         )
         context.argument(
@@ -502,8 +503,8 @@ def load_deviceupdate_arguments(self, _):
             "A --file entry can include the closest --related-file entries if provided. "
             "The following keys are supported: "
             "`path` [required] local file path to update file, "
-            "`downloadHandler` (ex: 'microsoft/delta:1') download handler for utilizing related files "
-            "to download payload file. "
+            "`downloadHandler` (ex: 'microsoft/delta:1') handler for utilizing related files to download payload file, "
+            "`properties` (in-line json object the agent will pass to the handler). "
             "--file can be used 1 or more times."
         )
         context.argument(
@@ -515,6 +516,21 @@ def load_deviceupdate_arguments(self, _):
             "A --related-file entry will be associated to the closest --file entry if it exists. "
             "The following keys are supported: "
             "`path` [required] local file path to related update file, "
-            "`properties` in-line json object passed to the download handler. "
+            "`properties` (in-line json object passed to the download handler). "
             "--related-file can be used 1 or more times."
+        )
+        context.argument(
+            "file_paths",
+            options_list=["--file-path"],
+            nargs="+",
+            action="append",
+            help="Local path to target file for hash calculation. "
+            "--file-path can be used 1 or more times."
+        )
+        context.argument(
+            "hash_algo",
+            options_list=["--hash-algo"],
+            help="Cryptographic algorithm to use for hashing.",
+            arg_type=get_enum_type(ADUValidHashAlgorithmType),
+            type=str,
         )