+++ azext_iot/tests/deviceupdate/test_adu_manifest_int.py
@@ -4,14 +4,21 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
+import os
 import pytest
+import json
+from pathlib import PurePath
+from typing import List
 from knack.log import get_logger
 from azext_iot.common.embedded_cli import EmbeddedCLI
-from azext_iot.tests.conftest import get_context_path
+from azext_iot.tests.conftest import get_context_path, generate_generic_id
 
 cli = EmbeddedCLI()
 
 logger = get_logger(__name__)
+
+unique_description = f"{generate_generic_id()} {generate_generic_id()}"
+unique_properties = {generate_generic_id(): generate_generic_id(), generate_generic_id(): unique_description}
 
 
 @pytest.mark.parametrize(
@@ -38,6 +45,62 @@ logger = get_logger(__name__)
                         "sizeInBytes": 163,
                         "hashes": {"sha256": "iFWTIaxp33tf5BR1w0fMmnnHpjsUjLRQ9eZFjw74LbU="},
                     }
+                ],
+                "manifestVersion": "5.0",
+            },
+        ),
+        (
+            "--update-provider digimaun0 --update-name simplescriptupdate --update-version 1 "
+            f"--description '{unique_description}' "
+            "--compat manufacturer=Contoso model=Vacuum "
+            f"--step handler=microsoft/script:1 description='{unique_description}' "
+            f"--file path=\"{get_context_path(__file__, 'manifests', 'libcurl4-doc-apt-manifest.json')}\" "
+            f"properties='{json.dumps(unique_properties)}' "
+            f"--related-file path=\"{get_context_path(__file__, 'manifests', 'simple_apt_manifest_v5.json')}\" "
+            f"properties='{json.dumps(unique_properties)}' "
+            f"--related-file path=\"{get_context_path(__file__, 'manifests', 'surface15', 'parent.importmanifest.json')}\" "
+            f"--file path=\"{get_context_path(__file__, 'manifests', 'surface15', 'action.sh')}\" ",
+            {
+                "updateId": {"provider": "digimaun0", "name": "simplescriptupdate", "version": "1"},
+                "description": unique_description,
+                "compatibility": [
+                    {"manufacturer": "Contoso", "model": "Vacuum"},
+                ],
+                "instructions": {
+                    "steps": [
+                        {
+                            "handler": "microsoft/script:1",
+                            "files": ["libcurl4-doc-apt-manifest.json", "action.sh"],
+                            "type": "inline",
+                            "description": unique_description,
+                        }
+                    ]
+                },
+                "files": [
+                    {
+                        "filename": "libcurl4-doc-apt-manifest.json",
+                        "sizeInBytes": 163,
+                        "hashes": {"sha256": "iFWTIaxp33tf5BR1w0fMmnnHpjsUjLRQ9eZFjw74LbU="},
+                        "properties": unique_properties,
+                        "relatedFiles": [
+                            {
+                                "filename": "simple_apt_manifest_v5.json",
+                                "sizeInBytes": 1031,
+                                "hashes": {"sha256": "L+ZKmOOT3xRfHsFK7pcTXBLjeI2OFCW0855qIcV5sts="},
+                                "properties": unique_properties,
+                            },
+                            {
+                                "filename": "parent.importmanifest.json",
+                                "sizeInBytes": 1390,
+                                "hashes": {"sha256": "hos1UvCk66WmtL/SPNUmub+k302BM4gtWYtAF7tOCb4="},
+                            },
+                        ],
+                    },
+                    {
+                        "filename": "action.sh",
+                        "sizeInBytes": 33,
+                        "hashes": {"sha256": "n+KGjLjSGr7LVKsgWiExUDeU6Z2ZTJu0tpAWxkmYKxA="},
+                    },
                 ],
                 "manifestVersion": "5.0",
             },
@@ -162,7 +225,7 @@ def test_adu_manifest_init_v5(options, expected):
             "--update-provider digimaun --update-name invalid --update-version 1.0.0 "
             "--compat deviceManufacturer=Contoso deviceModel=Vacuum "
             "--step handler=microsoft/apt:1 "
-            "--file downhandler=abcd/123",
+            "--file downhandler=abcd/123"
         ),
         (
             # path key is required for --related-file
@@ -184,3 +247,40 @@ def test_adu_manifest_init_v5(options, expected):
 )
 def test_adu_manifest_init_v5_invalid_path_required(options):
     assert not cli.invoke(f"iot device-update update init v5 {options}").success()
+
+
+@pytest.mark.parametrize(
+    "files_count, expected_bytes",
+    [
+        (1, 1024),
+        (2, 512),
+        (3, 256),
+    ],
+)
+def test_adu_manifest_init_calculate_hash(files_count, expected_bytes):
+    from azext_iot.deviceupdate.providers.base import DeviceUpdateDataManager, FileMetadata
+
+    normalized_paths: List[PurePath] = []
+    metadata: List[FileMetadata] = []
+    for i in range(files_count):
+        content_file_path = PurePath(get_context_path(__file__, "manifests", generate_generic_id()))
+        target_bytes = os.urandom(expected_bytes)
+        with open(content_file_path, "wb") as f:
+            f.write(target_bytes)
+        normalized_paths.append(content_file_path)
+        metadata.append(DeviceUpdateDataManager.calculate_file_metadata(content_file_path))
+
+    cli_path_input = ""
+    for p in normalized_paths:
+        cli_path_input = cli_path_input + f" --file-path '{str(p)}'"
+
+    result = cli.invoke(f"iot device-update update init calculate-hash {cli_path_input}").as_json()
+
+    for i in range(files_count):
+        assert result[i]["hashAlgorithm"] == "sha256"
+        assert result[i]["uri"] == normalized_paths[i].as_uri()
+        assert result[i]["hash"] == metadata[i].hash
+        assert result[i]["bytes"] == metadata[i].bytes
+
+        if os.path.exists(normalized_paths[i]):
+            os.remove(normalized_paths[i])